refactor(agents): split persona_runtime.py into package
Convert the 979-line agents/persona_runtime.py module into an

agents/persona_runtime/ package with four submodules:

- __init__.py (311 lines): _LLMPersonaAgent class, on_event/on_tick,

  system prompt assembly, event formatting, _coerce_event_timeout

- action_loop.py (436 lines): _ActionLoopMixin with multi-turn

  tool-use loop, tool execution, action validation and parsing

- memory_context.py (277 lines): _MemoryContextMixin with memory

  context injection, _truncate_with_ellipsis, truncation constants

- state_persistence.py (119 lines): _StatePersistenceMixin with

  state serialization and memory lifecycle management

All existing import paths (from agents.persona_runtime import ...)

resolve unchanged. No behavioral changes. All 853 unit tests pass.

v0.2 release prep PR 1 — file-size hygiene.
