+++ agents/persona_runtime/state_persistence.py
@@ -1,0 +1,119 @@
+"""State persistence and memory lifecycle for _LLMPersonaAgent.
+
+Handles serializing/deserializing persona state to the agent_state
+table and managing the initialization/shutdown of all three memory tiers.
+"""
+
+from __future__ import annotations
+
+import json
+import logging
+from typing import Any
+
+from ..memory.episodic import EpisodicMemory
+from ..memory.relationship import RelationshipMemory
+from ..memory.working import WorkingMemory
+from ..persona_types import PersonaState
+
+logger = logging.getLogger(__name__)
+
+__all__ = ["_StatePersistenceMixin"]
+
+
+class _StatePersistenceMixin:
+    """Mixin providing state persistence and memory lifecycle for _LLMPersonaAgent.
+
+    Expects the following attributes on ``self`` (provided by the
+    concrete ``_LLMPersonaAgent`` class and ``PersonaAgent`` base):
+
+    - ``agent_id: str``
+    - ``config: dict[str, Any]``
+    - ``_state: PersonaState``
+    - ``_episodic_memory: EpisodicMemory``
+    - ``_relationship_memory: RelationshipMemory``
+    - ``_working_memory: WorkingMemory``
+    - ``_lock: asyncio.Lock``
+    """
+
+    # Attribute declarations for type checkers — set by __init__.
+    agent_id: str
+    config: dict[str, Any]
+    _state: PersonaState
+    _episodic_memory: EpisodicMemory
+    _relationship_memory: RelationshipMemory
+    _working_memory: WorkingMemory
+    _lock: Any  # asyncio.Lock — avoid import for type annotation only
+
+    # ─── State persistence ─────────────────────────────
+
+    async def _persist_persona_state(self) -> None:
+        """Serialize persona state to the agent_state table.
+
+        Uses EpisodicMemory's public ``persist_agent_state()`` API rather
+        than reaching into its private DB handle (review finding #3).
+        """
+        try:
+            state_json = json.dumps(self._state.to_dict())
+            await self._episodic_memory.persist_agent_state(
+                self.agent_id, state_json,
+            )
+        except Exception:
+            logger.warning(
+                "Failed to persist persona state for agent %s",
+                self.agent_id,
+                exc_info=True,
+            )
+
+    async def _load_persona_state(self) -> PersonaState:
+        """Load persona state from the agent_state table, or return defaults.
+
+        Uses EpisodicMemory's public ``load_agent_state()`` API.
+        """
+        try:
+            state_json = await self._episodic_memory.load_agent_state(
+                self.agent_id,
+            )
+            if state_json:
+                return PersonaState.from_dict(json.loads(state_json))
+        except Exception:
+            logger.warning(
+                "Failed to load persona state for agent %s, using defaults",
+                self.agent_id,
+                exc_info=True,
+            )
+        return PersonaState()
+
+    # ─── Memory lifecycle ──────────────────────────────
+
+    async def initialize_memory(self) -> None:
+        """Initialize all memory tiers and load persisted state."""
+        await self._episodic_memory.initialize()
+        await self._relationship_memory.initialize(
+            config_relationships=self.config.get("relationships"),
+        )
+        await self._working_memory.initialize()
+        self._state = await self._load_persona_state()
+
+    async def close_memory(self) -> None:
+        """Close all memory tiers, awaiting in-flight operations.
+
+        Each tier is closed in its own try/except so that a failure in one
+        tier (e.g. disk-full on SQLite) does not prevent the remaining
+        tiers from releasing their resources (PR #54 review).
+        """
+        async with self._lock:
+            await self._persist_persona_state()
+            errors: list[Exception] = []
+            # Close order: working (flush compression) → episodic (DB) → relationship (DB)
+            for tier in (self._working_memory, self._episodic_memory, self._relationship_memory):
+                try:
+                    await tier.close()
+                except Exception as exc:
+                    errors.append(exc)
+                    logger.warning("Failed to close memory tier: %s", exc)
+            if errors:
+                logger.error(
+                    "Memory close for agent %s completed with %d error(s)",
+                    self.agent_id,
+                    len(errors),
+                )
